BUG Found! Input Eaten in PhysicsProcess()
- This bug was originally found when trying to dodge and nothing
happening (especially when trying to do it in the air when the player is
just about to hit the gound??). After some testing by adding the same
key/button detection logic in Process() and input methods, instead of
PhysicsProcess() the input was correctly registered in those functions.
- This leads me to conclude that the main issue is checking for
IsActionJustPressed() in PhysicsProcess() is unreliable and should
instead be checked in the proper UnhandledInput() Godot function (The
callback in my player state machine is HandleInput() or
HandleKeyboardInput())
- NOTE: It seems we can also check for the input in Process() reliably,
but in the feature we will eventually have to check for gui/menu input
and differentiate it form actual player movement input for example, and
the purpose of Godot's UnhandledInput() seems to be to help with this
kind of scenario.